+++ scape/core/src/main/java/me/scape/ti/dataobject/DesignContestUserDO.java
@@ -24,130 +24,130 @@ import javax.persistence.TemporalType;
 @Table(name = "design_contest_user", catalog = "scape")
 public class DesignContestUserDO implements Serializable {
 
-	private static final long serialVersionUID = -7782042921818978205L;
+    private static final long serialVersionUID = -7782042921818978205L;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	@Column(name = "id", unique = true, nullable = false)
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id", unique = true, nullable = false)
+    private Long id;
 
-	@Column(name = "contest_id", nullable = false)
-	private Long contest_id;// 赛事ID
+    @Column(name = "contest_id", nullable = false)
+    private Integer contest_id;// 赛事ID
 
-	@Column(name = "user_id", nullable = false)
-	private Long user_id;// 参赛者
+    @Column(name = "user_id", nullable = false)
+    private Integer user_id;// 参赛者
 
-	@Column(name = "number", unique = true, nullable = false, length = 64)
-	private String number;// 参赛编号
+    @Column(name = "number", unique = true, nullable = false, length = 64)
+    private String number;// 参赛编号
 
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "gmt_created", nullable = false)
-	private Date gmt_created;
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "gmt_created", nullable = false)
+    private Date gmt_created;
 
-	@Temporal(TemporalType.TIMESTAMP)
-	@Column(name = "gmt_modified", nullable = true)
-	private Date gmt_modified;
+    @Temporal(TemporalType.TIMESTAMP)
+    @Column(name = "gmt_modified", nullable = true)
+    private Date gmt_modified;
 
-	public Long getId() {
-		return id;
-	}
+    public Long getId() {
+        return id;
+    }
 
-	public void setId(Long id) {
-		this.id = id;
-	}
+    public void setId(Long id) {
+        this.id = id;
+    }
 
-	public Long getContest_id() {
-		return contest_id;
-	}
+    public Integer getContest_id() {
+        return contest_id;
+    }
 
-	public void setContest_id(Long contest_id) {
-		this.contest_id = contest_id;
-	}
+    public void setContest_id(Integer contest_id) {
+        this.contest_id = contest_id;
+    }
 
-	public Long getUser_id() {
-		return user_id;
-	}
+    public Integer getUser_id() {
+        return user_id;
+    }
 
-	public void setUser_id(Long user_id) {
-		this.user_id = user_id;
-	}
+    public void setUser_id(Integer user_id) {
+        this.user_id = user_id;
+    }
 
-	public String getNumber() {
-		return number;
-	}
+    public String getNumber() {
+        return number;
+    }
 
-	public void setNumber(String number) {
-		this.number = number;
-	}
+    public void setNumber(String number) {
+        this.number = number;
+    }
 
-	public Date getGmt_created() {
-		return gmt_created;
-	}
+    public Date getGmt_created() {
+        return gmt_created;
+    }
 
-	public void setGmt_created(Date gmt_created) {
-		this.gmt_created = gmt_created;
-	}
+    public void setGmt_created(Date gmt_created) {
+        this.gmt_created = gmt_created;
+    }
 
-	public Date getGmt_modified() {
-		return gmt_modified;
-	}
+    public Date getGmt_modified() {
+        return gmt_modified;
+    }
 
-	public void setGmt_modified(Date gmt_modified) {
-		this.gmt_modified = gmt_modified;
-	}
+    public void setGmt_modified(Date gmt_modified) {
+        this.gmt_modified = gmt_modified;
+    }
 
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((contest_id == null) ? 0 : contest_id.hashCode());
-		result = prime * result + ((gmt_created == null) ? 0 : gmt_created.hashCode());
-		result = prime * result + ((gmt_modified == null) ? 0 : gmt_modified.hashCode());
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		result = prime * result + ((number == null) ? 0 : number.hashCode());
-		result = prime * result + ((user_id == null) ? 0 : user_id.hashCode());
-		return result;
-	}
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((contest_id == null) ? 0 : contest_id.hashCode());
+        result = prime * result + ((gmt_created == null) ? 0 : gmt_created.hashCode());
+        result = prime * result + ((gmt_modified == null) ? 0 : gmt_modified.hashCode());
+        result = prime * result + ((id == null) ? 0 : id.hashCode());
+        result = prime * result + ((number == null) ? 0 : number.hashCode());
+        result = prime * result + ((user_id == null) ? 0 : user_id.hashCode());
+        return result;
+    }
 
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		DesignContestUserDO other = (DesignContestUserDO) obj;
-		if (contest_id == null) {
-			if (other.contest_id != null)
-				return false;
-		} else if (!contest_id.equals(other.contest_id))
-			return false;
-		if (gmt_created == null) {
-			if (other.gmt_created != null)
-				return false;
-		} else if (!gmt_created.equals(other.gmt_created))
-			return false;
-		if (gmt_modified == null) {
-			if (other.gmt_modified != null)
-				return false;
-		} else if (!gmt_modified.equals(other.gmt_modified))
-			return false;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		if (number == null) {
-			if (other.number != null)
-				return false;
-		} else if (!number.equals(other.number))
-			return false;
-		if (user_id == null) {
-			if (other.user_id != null)
-				return false;
-		} else if (!user_id.equals(other.user_id))
-			return false;
-		return true;
-	}
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        DesignContestUserDO other = (DesignContestUserDO) obj;
+        if (contest_id == null) {
+            if (other.contest_id != null)
+                return false;
+        } else if (!contest_id.equals(other.contest_id))
+            return false;
+        if (gmt_created == null) {
+            if (other.gmt_created != null)
+                return false;
+        } else if (!gmt_created.equals(other.gmt_created))
+            return false;
+        if (gmt_modified == null) {
+            if (other.gmt_modified != null)
+                return false;
+        } else if (!gmt_modified.equals(other.gmt_modified))
+            return false;
+        if (id == null) {
+            if (other.id != null)
+                return false;
+        } else if (!id.equals(other.id))
+            return false;
+        if (number == null) {
+            if (other.number != null)
+                return false;
+        } else if (!number.equals(other.number))
+            return false;
+        if (user_id == null) {
+            if (other.user_id != null)
+                return false;
+        } else if (!user_id.equals(other.user_id))
+            return false;
+        return true;
+    }
 }
